clean-up: EffectsProvider loads effects from knownPluginsRegister
