+++ src/effects/nyquist/internal/nyquisteffectsrepository.h
@@ -15,16 +15,17 @@
 namespace au::effects {
 class NyquistEffectsRepository : public INyquistEffectsRepository, public muse::Contextable
 {
+    muse::GlobalInject<muse::audioplugins::IKnownAudioPluginsRegister> knownPlugins;
     muse::GlobalInject<spectrogram::ISpectralEffectsRegister> spectralEffectsRegister;
 
 public:
     NyquistEffectsRepository(const muse::modularity::ContextPtr& ctx);
 
-    EffectMetaList effectMetaList() const override;
+    void init();
     bool ensurePluginIsLoaded(const EffectId& effectId) const override;
 
 private:
-    void registerSpectralEffects() const;
+    EffectMetaList effectMetaList() const;
 
     // This member forces the linker to include LoadNyquist.cpp,
     // which contains DECLARE_BUILTIN_PROVIDER for the Nyquist module
