+++ src/effects/nyquist/inyquisteffectsrepository.h
@@ -14,7 +14,6 @@ class INyquistEffectsRepository : MODULE_GLOBAL_INTERFACE
 public:
     virtual ~INyquistEffectsRepository() = default;
 
-    virtual EffectMetaList effectMetaList() const = 0;
     virtual bool ensurePluginIsLoaded(const EffectId& effectId) const = 0;
 };
 }
